third_party/blink/web_tests/external/wpt/web-animations/animation-model/animation-types/addition-per-property-001-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] font-style: "oblique" onto "italic"
  assert_equals: The value should be oblique at 0ms expected "oblique" but got "italic"
Harness: the test ran to completion.
